@@ -9,7 +9,7 @@ import Foundation
 
 /// An object that determines if an image is an individual image or a combined image
 /// comprised of individual images stitched together.
-public enum ImageType: String, CaseIterable, Identifiable {
+public enum ImageType: String, CaseIterable, Identifiable, Sendable {
     case individual = "Individual"
     case combined = "Combined"
     
