@@ -14,10 +14,10 @@ import OSLog
 import SBFoundation
 
 public struct CreateFramedScreenshotsIntent: AppIntent {
-    static let intentClassName = "CreateFramedScreenshotsIntent"
-    public static var title: LocalizedStringResource = "Create Framed Screenshots"
-    static var description = IntentDescription("Creates framed screenshots with a device frame using the images passed in.")
-    public static var isDiscoverable: Bool = true
+    public static let intentClassName = "CreateFramedScreenshotsIntent"
+    public static let title: LocalizedStringResource = "Create Framed Screenshots"
+    public static let description = IntentDescription("Creates framed screenshots with a device frame using the images passed in.")
+    public static let isDiscoverable: Bool = true
     private let logger = Logger(category: CreateFramedScreenshotsIntent.self)
     
     public init() { }
@@ -61,9 +61,9 @@ public struct CreateFramedScreenshotsIntent: AppIntent {
     // MARK: - Functions
     
     public func perform() async throws -> some IntentResult & ReturnsValue<[IntentFile]> {
-        let persistenceManager = PersistenceManager.shared
+        let persistenceManager = await PersistenceManager.shared
         
-        guard persistenceManager.canSaveFramedScreenshot else {
+        guard await persistenceManager.canSaveFramedScreenshot else {
             logger.error("pro subscription required to save screenshot")
             throw SBError.proSubscriptionRequired
         }
@@ -77,7 +77,9 @@ public struct CreateFramedScreenshotsIntent: AppIntent {
             return file
         }
         
-        persistenceManager.deviceFrameCreations += screenshots.count
+        await MainActor.run {
+            persistenceManager.deviceFrameCreations += screenshots.count
+        }
         
         logger.debug("returning CreateFramedScreenshotsIntent result")
         return .result(value: screenshots)
