@@ -9,7 +9,7 @@ import Foundation
 import AppIntents
 
 struct ShotBotShortcuts: AppShortcutsProvider {
-    static var shortcutTileColor: ShortcutTileColor = .blue
+    static let shortcutTileColor: ShortcutTileColor = .blue
     static var appShortcuts: [AppShortcut] {
         AppShortcut(
             intent: CreateFramedScreenshotsIntent(),
