@@ -109,7 +109,7 @@ public struct ImageManager: ImageManaging {
         }
         
         // Request the image
-        let (image, _) = await client.requestImage(
+        let image = await client.requestImage(
             latestScreenshotAsset,
             targetSize,
             .aspectFit,
@@ -159,7 +159,7 @@ public struct ImageManager: ImageManaging {
         let result = PHAsset.fetchAssets(with: .image, options: fetchOptions)
         
         let images = await result.phAssets.asyncCompactMap { asset in
-            let (image, _) = await client.requestImage(
+            let image = await client.requestImage(
                 asset,
                 .init(
                     width: asset.pixelWidth,
